feat: wip - process-zip endpoint
