+++ containers/ecr-viewer/src/app/api/process-zip/route.ts
@@ -1,6 +1,6 @@
-import { Bundle, FhirResource } from "fhir/r4";
-import { NextRequest, NextResponse } from "next/server";
 import { z } from "zod";
+
+import { postOrchestration } from "@/app/api/orchestration-utils";
 
 import { processZip } from "./service";
 
@@ -16,53 +16,23 @@ const schema = z.object({
     .transform((v) => v?.toLowerCase()),
 });
 
-interface ProcessZipResponse {
-  message: string;
-  erorors?: string[];
-  bundle?: Bundle<FhirResource>;
-}
-
 /**
  * Handles POST requests and saves the FHIR Bundle to the database.
  * @param request - The incoming request object.
  * @returns A `NextResponse` object with a JSON payload indicating the success message.
  */
-export async function POST(
-  request: NextRequest,
-): Promise<NextResponse<ProcessZipResponse>> {
-  // Parse out the form from the request
-  let formData: FormData;
-  try {
-    formData = await request.formData();
-  } catch {
-    return NextResponse.json(
-      { message: "Validation error", errors: ["No form found"] },
-      { status: 400 },
-    );
-  }
-
-  try {
-    const body = schema.parse(Object.fromEntries(formData));
-    const { status, ...payload } = await processZip(
-      body.upload_file,
-      body.return_fhir_bundle === "true",
-    );
-    return NextResponse.json(payload, { status });
-  } catch (error: unknown) {
-    if (error instanceof z.ZodError) {
-      return NextResponse.json(
-        {
-          message: "Validation error",
-          errors: error.errors,
-        },
-        { status: 400 },
-      );
-    }
-
-    console.error(error);
-    return NextResponse.json(
-      { message: "Internal Server Error" },
-      { status: 500 },
-    );
-  }
-}
+export const POST = postOrchestration(async (formData: FormData) => {
+  const body = schema.parse(Object.fromEntries(formData));
+  return await processZip(body.upload_file, body.return_fhir_bundle === "true");
+});
+// export async function POST(
+//   request: NextRequest,
+// ): Promise<NextResponse<ProcessOrchestrationResponse>> {
+//   return await postOrchestration(request, async (formData: FormData) => {
+//     const body = schema.parse(Object.fromEntries(formData));
+//     return await processZip(
+//       body.upload_file,
+//       body.return_fhir_bundle === "true",
+//     );
+//   });
+// }
